battery-status/battery-disabled-by-feature-policy.https.sub.html

<!DOCTYPE html>
<body>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/feature-policy/resources/featurepolicy.js></script>
<script>
'use strict';

let same_origin_src = '/feature-policy/resources/feature-policy-battery.html';
let cross_origin_src = 'https://{{domains[www]}}:{{ports[https][0]}}' +
  same_origin_src;
let header = 'Feature-Policy header {"battery" : []}';

promise_test(() => {
  return navigator.getBattery().then(() => {
    assert_unreached('expected promise to reject with SecurityError');
  }, error => {
    assert_equals(error.name, 'SecurityError');
  });
}, header + ' disallows the top-level document.');

async_test(t => {
  test_feature_availability('getBattery()', t, same_origin_src,
      expect_feature_unavailable_default);
}, header + ' disallows same-origin iframes.');

async_test(t => {
  test_feature_availability('getBattery()', t, cross_origin_src,
      expect_feature_unavailable_default);
}, header + ' disallows cross-origin iframes.');
</script>
</body>
